feat(products-management): better error handling

+++ libs/contexts/products-management/src/lib/products/application/handlers/delete-product.handler.ts
@@ -1,3 +1,8 @@
+import {
+  BaseError,
+  UnexpectedError,
+  DatabaseRecordNotFoundError,
+} from '@ecommerce-monorepo/shared';
 import { ProductRepository, ProductAggregate } from '../../domain';
 import { DeleteProductCommand } from '../commands/delete-product.command';
 
@@ -5,24 +10,26 @@ export class DeleteProductCommandHandler {
   constructor(private productRepository: ProductRepository) {}
 
   async execute(deleteProductCommand: DeleteProductCommand): Promise<void> {
-    let productAggregate: ProductAggregate;
-
     try {
-      productAggregate = ProductAggregate.fromPrimitives(
-        await this.productRepository.getProduct(deleteProductCommand.id),
+      const productPrimitive = await this.productRepository.getProduct(
+        deleteProductCommand.id,
       );
-    } catch (err) {
-      //TODO: Error más semántico
-      throw new Error();
-    }
 
-    try {
+      if (!productPrimitive)
+        throw new DatabaseRecordNotFoundError(this.constructor.name, 'execute');
+
+      ProductAggregate.fromPrimitives(productPrimitive);
+
       return await this.productRepository.deleteProduct(
-        productAggregate.toPrimitives().id,
+        deleteProductCommand.id,
       );
     } catch (err) {
-      //TODO: Error más semántico
-      throw new Error();
+      if (err instanceof BaseError) throw err;
+      throw new UnexpectedError(
+        this.constructor.name,
+        'execute',
+        JSON.stringify(deleteProductCommand),
+      );
     }
   }
 }
